requirements.txt
asgiref==3.8.1
Django==5.2.1
django-vite==3.1.0
sqlparse==0.5.3
django-allauth[socialaccount]
